Reorganize header file
