+++ include/dgl/shared_mem_manager.h
