+++ src/array/array_arith.cc
@@ -6,7 +6,7 @@
 #include <dgl/packed_func_ext.h>
 #include <dgl/runtime/ndarray.h>
 #include <dgl/runtime/container.h>
-#include "../c_api_common.h"
+#include <dgl/c_api_common.h>
 #include "./array_op.h"
 #include "./arith.h"
