+++ src/graph/graph_traversal.cc
@@ -5,7 +5,7 @@
  */
 #include <dgl/graph_traversal.h>
 #include <dgl/packed_func_ext.h>
-#include "../c_api_common.h"
+#include <dgl/c_api_common.h>
 
 using namespace dgl::runtime;
