+++ src/graph/immutable_graph.cc
@@ -8,6 +8,8 @@
 #include <dgl/packed_func_ext.h>
 #include <dgl/runtime/smart_ptr_serializer.h>
 #include <dgl/base_heterograph.h>
+#include <dgl/heterograph.h>
+#include <dgl/unit_graph.h>
 #include <dmlc/io.h>
 #include <dmlc/type_traits.h>
 #include <string.h>
@@ -15,9 +17,7 @@
 #include <numeric>
 #include <tuple>
 
-#include "../c_api_common.h"
-#include "heterograph.h"
-#include "unit_graph.h"
+#include <dgl/c_api_common.h>
 
 using namespace dgl::runtime;
