+++ src/graph/subgraph.cc
@@ -3,7 +3,7 @@
  * \file graph/subgraph.cc
  * \brief Functions for extracting subgraphs.
  */
-#include "./heterograph.h"
+#include <dgl/heterograph.h>
 using namespace dgl::runtime;
 
 namespace dgl {
