+++ src/graph/transform/metis_partition_hetero.cc
@@ -6,10 +6,10 @@
 
 #include <dgl/base_heterograph.h>
 #include <dgl/packed_func_ext.h>
+#include <dgl/unit_graph.h>
+#include <dgl/heterograph.h>
 #include <metis.h>
 
-#include "../heterograph.h"
-#include "../unit_graph.h"
 
 using namespace dgl::runtime;
